removed single day container and refactored to render from Moods, share state between week and day displays
