+++ src/components/WeekDisplay.js
@@ -11,34 +11,10 @@ import MoodPreview from './MoodPreview';
 import { MoodsPrevWeekContext } from '../contexts/MoodsPrevWeekContext';
 import styles from '../styles/theme';
 
-function WeekDisplay() {
-  const { user } = useAuth();
-
-  const { loading, error, data } = useQuery(checkForUserAndGetMoodsQuery, {
-    variables: {
-      email: user.email,
-      firstName: user.given_name,
-      lastName: user.family_name,
-    },
-  });
-
-  const { moods, setMoods } = useContext(MoodsPrevWeekContext);
-
-  // set moodsPrevWeek context if the data from query exists
-  useEffect(() => {
-    if (data) {
-      setMoods(weekOfMoods(data.user.moods));
-    }
-  }, [data, setMoods]);
-
-  if (error) return <p>{error.message}</p>;
-
-  return loading ? (
-    <LoadingWrapper>
-      <Spin size="large" delay={300} />
-    </LoadingWrapper>
-  ) : (
+function WeekDisplay({ moods, handleCurrentDay }) {
+  return (
     <>
+      <StyledLogo>MoodBloom</StyledLogo>
       <Greeting>Weekly moods</Greeting>
       {moods &&
         moods.map((list) => {
@@ -46,8 +22,9 @@ function WeekDisplay() {
           if (list.length !== 0) {
             return (
               <Link
-                to={`/dashboard/day/${getDay(+list[0].createdAt)}`}
+                to="/dashboard/day"
                 key={list[0].id}
+                onClick={() => handleCurrentDay(getDay(+list[0].createdAt))}
               >
                 <MoodPreview
                   count={list.length}
@@ -72,13 +49,10 @@ const Greeting = styled.h2`
   margin-bottom: 30px;
 `;
 
-const LoadingWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 50%;
-
-  .ant-spin-dot-item {
-    background-color: ${styles.darkJungleGreen} !important;
-  }
+const StyledLogo = styled.h1`
+  font-size: 21px;
+  font-weight: 600;
+  color: #0c423b;
+  margin: unset;
+  margin-bottom: 31px;
 `;
