+++ src/containers/App/App.js
@@ -16,7 +16,6 @@ import EntryForm from '../EntryForm/EntryForm';
 import Moods from '../Moods';
 import NotFound from '../NotFound/404';
 import Settings from '../Settings';
-import SingleDay from '../SingleDay';
 import { MoodsPrevWeekProvider } from '../../contexts/MoodsPrevWeekContext';
 import styles from '../../styles/theme';
 
@@ -48,9 +47,8 @@ function App() {
             <Route path="/signin" exact component={SignIn} />
             <Route path="/signup" exact component={SignUp} />
             <PrivateRoute path="/entryform" component={EntryForm} />
-            <PrivateRoute path="/dashboard" exact component={Moods} />
-            <PrivateRoute path="/dashboard/day/:day" component={SingleDay} />
             <PrivateRoute path="/dashboard/settings" component={Settings} />
+            <PrivateRoute path="/dashboard" component={Moods} />
             <PrivateRoute path="/profile" component={Profile} />
             <Route component={NotFound} />
           </Switch>
