--- src/containers/SingleDay/SingleDay.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import MoodDisplay from '../../components/MoodDisplay';
-
-function SingleDay() {
-  return <MoodDisplay />;
-}
-
-export default SingleDay;
